Add files via upload
The most popular starting stations are  near the public transportation centers such as train stations and bus stops. This is because riders tend to ride bikes to where they can take public transportations to either work or home .

This visualization on the top two slides shows the map and the streets where most rides are taking place. The bottom two charts slides show the top 10 starting and ending stations.

https://public.tableau.com/app/profile/elias.mwita/viz/PopularStationinJune2021/CitiBikeRideAnalysisforJune2021?publish=yes